Fix SPCS errors.
The state plane coordinate system parameters are stored in a binary
format. The files in the original archive are stored in big-endian
format. To avoid problems on some platforms we generate the binary files
from the ascii source files during the build.

Remove hard-coded paths from spload.f and pass input and output files on
the command-line to allow for VPATH builds.

Fix 2 bugs in the binary file reading/writing routines:

 - The 'id' field in the binary parameter files is a 4-byte integer, so
   we use 'int32_t' instead of 'long int'.

 - Change the record length parameter in spload.f to 432, in order to
   match the record length used in the reading routines in stplnfor.c
   and stplninv.c.