Adapt test code for use with Automake.
The original test suite required the user to manually run tests and
inspect their output. In order for 'make check' to produce something
meaningful, we need an automatic way to check if tests ran
successfully.

Replace the single test script Proj.name by individual calls to tests
for each projection.

grep the output files of each test to look for error messages.

Add a command line argument to testproj so we can pass the correct
location for the input parameter files.  This is necessary for the tests
to work with VPATH builds.

Move the test scripts out of the test/proj/* directories which are
distributed with EXTRA_DIST. Otherwise, test .log and .trs files end up
in the distributed tarball.

Remove close_file() from report.c to avoid double free errors (the
FILE's are always closed closed immediately).

Make SPCS param generation work with VPATH build.